Update and rename Area.java to SquareArea.java
